Moved objectClasses from app/database into server/database
