+++ server/database/objectClasses/Exhibition.js
